refactor: 데이터 레이어 패키지 정리
