+++ app/src/main/java/woowacourse/shopping/data/carts/CartDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.carts
 
 import androidx.room.Dao
 import androidx.room.Delete
