+++ app/src/main/java/woowacourse/shopping/data/carts/CartEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.carts
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
